Added .hide Class for Element Visibility Control
In this commit, a new CSS class named ".hide" has been introduced to the stylesheet. This class applies the CSS property "display: none;" to elements, effectively hiding them from view and removing them from the document flow. The ".hide" class will be used for controlling the visibility of specific elements in the HTML document, enhancing the user interface's interactive behavior.